pyproject.toml
[build-system]
requires = ["setuptools~=80.0.0", "wheel~=0.45.0"]
build-backend = "setuptools.build_meta"

[project]
name = "money-manager"
version = "0.1.0"
description = "Money Manager Application"
readme = "README.md"
requires-python = ">=3.13"
keywords = ["money", "expense", "budget", "finance"]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop"
]
dependencies = [
    "python-dotenv==1.2.1",
    "python-dateutil==2.9.0",
    "fastapi==0.128.0",
    "uvicorn[standard]==0.40.0",
    "streamlit==1.52.2",
]

[project.optional-dependencies]
dev = [
    "pytest~=9.0.0",
    "httpx~=0.28.1"
]

[project.scripts]
money-manager = "app.main:main"
money-manager-api = "app.api.main:run_server"
money-manager-web = "app.web.main:run_app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
